+++ config/filament-record-nav.php
@@ -1,5 +1,8 @@
 <?php
 
+use Filament\Support\Enums\IconPosition;
+use Filament\Support\Icons\Heroicon;
+
 return [
     /*
     |--------------------------------------------------------------------------
@@ -37,4 +40,35 @@ return [
     */
     'previous_direction' => 'desc',
     'next_direction'     => 'asc',
+
+    /*
+     |--------------------------------------------------------------------------
+     | Display Record Title
+     |--------------------------------------------------------------------------
+     | determine if the title should be displayed in the button
+     */
+    'display_record_title' => true,
+
+    /*
+    |--------------------------------------------------------------------------
+    | Icons Used
+    |--------------------------------------------------------------------------
+    | specified the icons used for next and previous buttons, you can override this
+    | global configuration by adding following attributes to the page:
+    |    public static $previousRecordIcon = Heroicon::SignalSlash;
+    |    public static $nextRecordIcon = Heroicon::CheckCircle;
+    | the value can be string|BackedEnum|Htmlable similar to resource icon
+    */
+    'previous_icon' => Heroicon::ChevronLeft,
+    'next_icon' => Heroicon::ChevronRight,
+
+    /*
+    |--------------------------------------------------------------------------
+    | Icons Position
+    |--------------------------------------------------------------------------
+    | specified the icon position it make different when the record title is enabled,
+    | the value can be string|BackedEnum referee to iconPosition @ filament doc.
+    */
+    'previous_icon_position' => IconPosition::After,
+    'next_icon_position' => IconPosition::Before,
 ];